+++ ZimLX/src/org/zimmob/zimlx/ZimIconLoader.kt
@@ -68,6 +68,6 @@ class ZimIconLoader(private val context: Context, iconCache: TaskKeyLruCache<Dra
         return mDrawableFactory.newIcon(bitmapInfo, activityInfo)
     }
 }
-*/
 
-//class NonAdaptiveIconDrawable(dr: Drawable?) : DrawableWrapper(dr)
+
+class NonAdaptiveIconDrawable(dr: Drawable?) : DrawableWrapper(dr)*/