+++ ZimLX/src/org/zimmob/zimlx/preferences/GridSizePreference.kt
@@ -14,6 +14,7 @@ class GridSizePreference(context: Context, attrs: AttributeSet?) :
 
     init {
         updateSummary()
+
     }
 
     fun getSize(): Pair<Int, Int> {