+++ src/com/android/launcher3/allapps/AllAppsStore.java
@@ -34,6 +34,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Set;
 
+import androidx.annotation.Nullable;
+
 /**
  * A utility class to maintain the collection of all apps.
  */
@@ -61,6 +63,16 @@ public class AllAppsStore {
 
     public AppInfo getApp(ComponentKey key) {
         return mComponentToAppMap.get(key);
+    }
+
+    @Nullable
+    public AppInfo getApp(String packageName) {
+        for (AppInfo app : mComponentToAppMap.values()) {
+            if (app.componentName.getPackageName().equals(packageName)) {
+                return app;
+            }
+        }
+        return null;
     }
 
     public void setDeferUpdates(boolean deferUpdates) {